задача 6 д/з заменила int на double

+++ Seminar_1/ДомашниеЗадания/ДЗ_3/Program.cs
@@ -1,7 +1,7 @@
 ﻿// Напишите программу, которая на вход принимает число и выдаёт, является ли число чётным (делится ли оно на два без остатка)
 
 Console.Write("Input a ");
-int a = Convert.ToInt32(Console.ReadLine());
+double a = Convert.ToInt32(Console.ReadLine());
 
 if(a%2 == 0)
 Console.WriteLine(a + "->" + "Yes");
